requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
pytz==2023.3.post1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
typing_extensions==4.8.0
urllib3==2.1.0
